config: extract location structs from builtin
Before reusing these concepts in builtin/config-batch.c, extract the
config_location_options struct from builtin/config.c to config.h with
implementation in config.c.

The only modification in this conversion is the use of a repository
parameter instead of the_repository.
